refactor: patient endpoints into routes/controller
